try to fix import issue #794
